# proto/wordsearch.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/wordsearch.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/words": {
      "get": {
        "operationId": "SingleWordSearch",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/wordsearchSingleWordSearchResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WordSearch"
        ]
      }
    },
    "/api/words/popular": {
      "get": {
        "operationId": "TopFiveSearchResults",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/wordsearchTopFiveResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "WordSearch"
        ]
      }
    },
    "/api/words/{word}": {
      "post": {
        "operationId": "UpdateSearchList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/wordsearchUpdateSearchListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "word",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/wordsearchUpdateSearchListRequest"
            }
          }
        ],
        "tags": [
          "WordSearch"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "wordsearchSearchItem": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "search_count": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "wordsearchSingleWordSearchResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "wordsearchTopFiveResponse": {
      "type": "object",
      "properties": {
        "top_five": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/wordsearchSearchItem"
          }
        }
      }
    },
    "wordsearchUpdateSearchListRequest": {
      "type": "object",
      "properties": {
        "word": {
          "type": "string"
        }
      }
    },
    "wordsearchUpdateSearchListResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "word_list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/wordsearchSearchItem"
          }
        }
      }
    }
  }
}